refactor: move RootModule to new submodule
