+++ crates/prql-compiler/src/semantic/resolver/mod.rs
@@ -10,9 +10,9 @@ use crate::semantic::context::TableDecl;
 use crate::semantic::{static_analysis, NS_PARAM};
 use crate::utils::IdGenerator;
 
-use super::context::{Decl, DeclKind, RootModule, TableExpr};
+use super::context::{Decl, DeclKind, TableExpr};
 use super::module::Module;
-use super::{write_pl, NS_DEFAULT_DB, NS_INFER, NS_STD, NS_THAT, NS_THIS};
+use super::{write_pl, RootModule, NS_DEFAULT_DB, NS_INFER, NS_STD, NS_THAT, NS_THIS};
 use flatten::Flattener;
 use transforms::coerce_into_tuple_and_flatten;
 use type_resolver::infer_type;
